+++ tests/Support/SpyStreamObserver.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Anktx\Kafka\Client\Tests\Support;
 
-use Anktx\Kafka\Client\ConsumeResult\KafkaBrokersDown;
 use Anktx\Kafka\Client\ConsumeResult\KafkaConsumeTimeout;
 use Anktx\Kafka\Client\ConsumeResult\KafkaPartitionEof;
 use Anktx\Kafka\Client\KafkaMessage\KafkaConsumerMessage;
@@ -24,9 +23,6 @@ final class SpyStreamObserver implements StreamObserver
     /** @var list<KafkaConsumeTimeout> */
     public array $timeouts = [];
 
-    /** @var list<KafkaBrokersDown> */
-    public array $brokersDown = [];
-
     /** @var list<KafkaPartitionEof> */
     public array $eofs = [];
 
@@ -38,11 +34,6 @@ final class SpyStreamObserver implements StreamObserver
     public function onTimeout(KafkaConsumeTimeout $timeout): void
     {
         $this->timeouts[] = $timeout;
-    }
-
-    public function onBrokersDown(KafkaBrokersDown $brokersDown): void
-    {
-        $this->brokersDown[] = $brokersDown;
     }
 
     public function onEof(KafkaPartitionEof $eof): void